requirements.txt
# Machine Learning Compulsory
albumentations==1.3.0
datasets==3.2.0
joblib==1.4.2
matplotlib==3.10.0
monai==1.4.0
numpy==1.26.4
opencv-python-headless==4.11.0.86
pandas==2.2.3
patchify==0.2.3
pillow==11.0.0
pydantic==2.8.2
pydantic_core==2.20.1
scikit-learn==1.6.1
scipy==1.15.2
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==1.0.13
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
wandb==0.19.7
pytorch-lightning
vessl==0.1.191
transformers==4.39.0
easydict==1.10
seaborn

# not sure
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
boto3==1.37.4
botocore==1.37.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==1.6.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
durationpy==0.9
ecdsa==0.19.0
editor==1.6.6
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
halo==0.0.31
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
inquirer==3.4.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.8
kubernetes==32.0.1
log-symbols==0.0.14
matplotlib-inline==0.1.7
multidict==6.1.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mpmath==1.3.0
multiprocess==0.70.16
networkx==3.4.2
oauthlib==3.2.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
paramiko==3.5.1
parso==0.8.4
platformdirs==4.3.6
prometheus_client==0.20.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.1
requests==2.32.3
requests-futures==1.0.2
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
rsa==4.9
runs==1.2.2
s3transfer==0.11.3
safetensors==0.5.2
selenium==4.28.1
Send2Trash==1.8.3
sentry-sdk==2.22.0
setproctitle==1.3.5
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spinners==0.0.24
sshpubkeys==3.3.1
stack-data==0.6.3
stringzilla==3.12.0
tabulate==0.9.0
termcolor==2.5.0
terminaltables==3.1.10
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
trio==0.28.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==3.0.0
webcolors==24.11.1
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
xmod==1.8.1
xxhash==3.5.0
yarl==1.18.3